+++ src/d3d9/d3d9_util.h
@@ -8,6 +8,7 @@
 #include "../dxvk/dxvk_device.h"
 
 #include "../util/util_matrix.h"
+#include "../util/util_misc.h"
 
 #include <d3dcommon.h>
 
@@ -100,14 +101,6 @@ namespace dxvk {
   VkFormatFeatureFlags GetImageFormatFeatures(DWORD Usage);
 
   VkImageUsageFlags GetImageUsageFlags(DWORD Usage);
-
-  inline void DecodeD3DCOLOR(D3DCOLOR color, float* rgba) {
-    // Encoded in D3DCOLOR as argb
-    rgba[3] = (float)((color & 0xff000000) >> 24) / 255.0f;
-    rgba[0] = (float)((color & 0x00ff0000) >> 16) / 255.0f;
-    rgba[1] = (float)((color & 0x0000ff00) >> 8)  / 255.0f;
-    rgba[2] = (float)((color & 0x000000ff))       / 255.0f;
-  }
 
   inline VkFormat PickSRGB(VkFormat format, VkFormat srgbFormat, bool srgb) {
     return srgb ? srgbFormat : format;
